backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlmodel==0.0.24
httpx==0.28.1
feedparser==6.0.11
pydantic-settings==2.8.1
openai>=1.68.0
pytest==8.3.5
